Full migration from CommonJS to ES2015 module syntax. Use babel-node instead of node to make this work for node scripts. Also move from_frac from shipengine.js to ui-functions.js. (This could only be done without breaking scripts via babel-node.)
