+++ lib/namedvector.js
@@ -1,4 +1,3 @@
-
 class NamedVector {
 	constructor(val, shortnames) {
 		this.shortnames = shortnames;
@@ -82,4 +81,6 @@ class NamedVector {
 	};
 };
 
-module.exports.NamedVector = NamedVector;
+export {
+	NamedVector,
+}
