+++ lib/shipengine.js
@@ -1,5 +1,6 @@
-const NamedVector = require('./namedvector').NamedVector
-const _ = require('lodash');
+import { from_frac } from '../src/ui-functions.js';
+import { NamedVector } from './namedvector.js';
+import _ from 'lodash';
 
 // AL6, $AL$6
 const SIZE_TO_WEIGHT_CAP_MULTIPLIER = 300
@@ -1739,19 +1740,9 @@ class DB {
 	};
 };
 
-// TODO: move this to ui-functions.js once we figure out to get ES2015 module imports working here
-function from_frac(str) {
-	let m = /(\d+ )?(\d+)\/(\d+)/.exec(str);
-	if (!m) {
-		return null;
-	}
-	let whole = m[1] ? Number.parseInt(m[1]) : 0;
-	let numer = Number.parseInt(m[2]);
-	let denom = Number.parseInt(m[3]);
-	return whole + numer / denom;
+export {
+	Design,
+	DB,
+	Statline,
+	Crewline,
 }
-
-module.exports.Design = Design;
-module.exports.DB = DB;
-module.exports.Statline = Statline;
-module.exports.Crewline = Crewline;
