+++ lib/shipimporter.js
@@ -1,8 +1,6 @@
-const util = require('util')
-const NamedVector = require('./namedvector').NamedVector
-
-const Papa = require('papaparse');
-const ShipEngine = require('./shipengine');
+import { NamedVector } from './namedvector.js';
+import * as ShipEngine from './shipengine.js';
+import Papa from 'papaparse';
 
 const COMPONENT_MAPPING = [
 	{
@@ -111,4 +109,6 @@ function import_design(design_csv) {
 	};
 };
 
-module.exports.import_design = import_design;
+export {
+	import_design,
+}
