+++ package.json
@@ -1,6 +1,7 @@
 {
     "dependencies": {
         "argparse": "^1.0.10",
+        "babel-cli": "^6.26.0",
         "bluebird": "^3.5.1",
         "lodash": "^4.17.5",
         "papaparse": "^4.3.7",
@@ -8,9 +9,9 @@
         "vuex": "^3.0.1"
     },
     "scripts": {
-        "load-ship": "node scripts/load-ship.js --parts dist/parts_C8.csv  --frames dist/frames_C8.csv --modules dist/modules_C8.csv",
-        "import-ship": "node scripts/import-ship.js --parts dist/parts_C8.csv  --frames dist/frames_C8.csv --modules dist/modules_C8.csv",
-        "import-parts": "node scripts/import-parts.js",
+        "load-ship": "babel-node scripts/load-ship.js --parts dist/parts_C8.csv --frames dist/frames_C8.csv --modules dist/modules_C8.csv",
+        "import-ship": "babel-node scripts/import-ship.js --parts dist/parts_C8.csv --frames dist/frames_C8.csv --modules dist/modules_C8.csv",
+        "import-parts": "babel-node scripts/import-parts.js --parts dist/parts_C8.csv --frames dist/frames_C8.csv --modules dist/modules_C8.csv 'Starfleet Canon C8'",
         "webpack": "webpack",
         "watch": "webpack --progress --watch",
         "start": "webpack-dev-server",
