+++ src/component.vue
@@ -55,7 +55,7 @@
 
 <script>
 
-import ShipEngine from '../lib/shipengine.js';
+import * as ShipEngine from '../lib/shipengine.js';
 
 import {
 	pretty,
