+++ src/design-summary.vue
@@ -44,7 +44,7 @@
 
 import { mapState, mapGetters } from 'vuex';
 
-import ShipEngine from '../lib/shipengine.js';
+import * as ShipEngine from '../lib/shipengine.js';
 
 import { frac } from './ui-functions.js';
