+++ src/principal-frame-final.vue
@@ -37,7 +37,7 @@
 
 import { mapState, mapGetters } from 'vuex';
 
-import ShipEngine from '../lib/shipengine.js';
+import * as ShipEngine from '../lib/shipengine.js';
 
 import StatlineCell from './statline-cell.vue';
