+++ src/setting-cell.vue
@@ -20,7 +20,7 @@
 
 <script>
 
-import ShipEngine from '../lib/shipengine.js';
+import * as ShipEngine from '../lib/shipengine.js';
 
 export default {
 	name: 'SettingCell',
