+++ src/subsystem-summary.vue
@@ -36,7 +36,7 @@
 
 import { mapState, mapGetters } from 'vuex';
 
-import ShipEngine from '../lib/shipengine.js';
+import * as ShipEngine from '../lib/shipengine.js';
 
 import StatlineCell from './statline-cell.vue';
